meia bomba.
Funcionando os output, mas parou de funcionar os inputs. Entretanto, creio que é algo bobo

+++ HMS/HMS.Nugets/Nugets/Nuget.Contracts/Inputs/EmployeeContractInput.cs
@@ -18,7 +18,7 @@ namespace Nuget.Contracts.Inputs
         public short JobTitle { get; set; }// Simucação de um banco de dados dedicado
         public sbyte ExperienceLevel { get; set; } 
         public List<short> Benefits { get; set; } = new(); // Simulacão de beneficios que podem estar armazenados em um banco de dados
-        public int Department { get; set; } //Enum ou um banco de dados dedicado
+        public short Department { get; set; } //Enum ou um banco de dados dedicado
         public short HourlySalaryInDollar { get; set; } // Praticamente ninguem recebe +32k/hora
         
         public Guid WorkHoursID { get; set; }
